abstracting in controllers
